feat: ini transaction fetcher
