+++ packages/stores/src/index.ts
@@ -30,3 +30,8 @@ export { useSettings } from "./useSettings";
 export { useTables } from "./useTables";
 export type { PendingTransaction } from "./useTransactions";
 export { useTransactions } from "./useTransactions";
+export type {
+  UserOperationDetailsItem,
+  UserOperationDevInfo,
+} from "./useUserOperations";
+export { useUserOperations } from "./useUserOperations";
